Trello auth (#265)
* (wip)(no land) trello auth

* save

* save

* ready to test maybe

* save

* tests pass

+++ backend/api/authorize.go
@@ -2,13 +2,18 @@ package api
 
 import (
 	"github.com/GeneralTask/task-manager/backend/database"
+	"github.com/GeneralTask/task-manager/backend/external"
 
 	"github.com/gin-gonic/gin"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
-// RedirectParams ...
-type RedirectParams struct {
+type Oauth1RedirectParams struct {
+	Token    string `form:"oauth_token" binding:"required"`
+	Verifier string `form:"oauth_verifier" binding:"required"`
+}
+
+type Oauth2RedirectParams struct {
 	Code  string `form:"code" binding:"required"`
 	State string `form:"state" binding:"required"`
 }
@@ -23,21 +28,24 @@ func (api *API) Authorize(c *gin.Context) {
 	if err != nil {
 		return
 	}
-	db, dbCleanup, err := database.GetDBConnection()
-	if err != nil {
-		Handle500(c)
-		return
-	}
-	defer dbCleanup()
-	insertedStateToken, err := database.CreateStateToken(db, &internalToken.UserID)
-	if err != nil {
-		Handle500(c)
-		return
-	}
-	stateTokenID, err := primitive.ObjectIDFromHex(*insertedStateToken)
-	if err != nil {
-		Handle500(c)
-		return
+	stateTokenID := primitive.NilObjectID
+	if taskService.Details.AuthType == external.AuthTypeOauth2 {
+		db, dbCleanup, err := database.GetDBConnection()
+		if err != nil {
+			Handle500(c)
+			return
+		}
+		defer dbCleanup()
+		insertedStateToken, err := database.CreateStateToken(db, &internalToken.UserID)
+		if err != nil {
+			Handle500(c)
+			return
+		}
+		stateTokenID, err = primitive.ObjectIDFromHex(*insertedStateToken)
+		if err != nil {
+			Handle500(c)
+			return
+		}
 	}
 	authURL, err := taskService.Service.GetLinkURL(stateTokenID, internalToken.UserID)
 	if err != nil {
@@ -57,28 +65,39 @@ func (api *API) AuthorizeCallback(c *gin.Context) {
 	if err != nil {
 		return
 	}
-	var redirectParams RedirectParams
-	if c.ShouldBind(&redirectParams) != nil || redirectParams.Code == "" || redirectParams.State == "" {
-		c.JSON(400, gin.H{"detail": "Missing query params"})
-		return
+	callbackParams := external.CallbackParams{}
+	if taskServiceResult.Details.AuthType == external.AuthTypeOauth1 {
+		var redirectParams Oauth1RedirectParams
+		if c.ShouldBind(&redirectParams) != nil || redirectParams.Token == "" || redirectParams.Verifier == "" {
+			c.JSON(400, gin.H{"detail": "missing query params"})
+			return
+		}
+		callbackParams = external.CallbackParams{Oauth1Token: &redirectParams.Token, Oauth1Verifier: &redirectParams.Verifier}
+	} else if taskServiceResult.Details.AuthType == external.AuthTypeOauth2 {
+		var redirectParams Oauth2RedirectParams
+		if c.ShouldBind(&redirectParams) != nil || redirectParams.Code == "" || redirectParams.State == "" {
+			c.JSON(400, gin.H{"detail": "missing query params"})
+			return
+		}
+		stateTokenID, err := primitive.ObjectIDFromHex(redirectParams.State)
+		if err != nil {
+			c.JSON(400, gin.H{"detail": "invalid state token format"})
+			return
+		}
+		db, dbCleanup, err := database.GetDBConnection()
+		if err != nil {
+			Handle500(c)
+			return
+		}
+		defer dbCleanup()
+		err = database.DeleteStateToken(db, stateTokenID, &internalToken.UserID)
+		if err != nil {
+			c.JSON(400, gin.H{"detail": "invalid state token"})
+			return
+		}
+		callbackParams = external.CallbackParams{Oauth2Code: &redirectParams.Code}
 	}
-	stateTokenID, err := primitive.ObjectIDFromHex(redirectParams.State)
-	if err != nil {
-		c.JSON(400, gin.H{"detail": "invalid state token format"})
-		return
-	}
-	db, dbCleanup, err := database.GetDBConnection()
-	if err != nil {
-		Handle500(c)
-		return
-	}
-	defer dbCleanup()
-	err = database.DeleteStateToken(db, stateTokenID, &internalToken.UserID)
-	if err != nil {
-		c.JSON(400, gin.H{"detail": "invalid state token"})
-		return
-	}
-	err = taskServiceResult.Service.HandleLinkCallback(redirectParams.Code, internalToken.UserID)
+	err = taskServiceResult.Service.HandleLinkCallback(callbackParams, internalToken.UserID)
 	if err != nil {
 		c.JSON(500, gin.H{"detail": err.Error()})
 		return
